requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.9.0
opencv-contrib-python>=4.9.0
scikit-learn>=1.0.0
pyyaml>=6.0
pandas>=1.3.0
matplotlib>=3.5.0
tqdm>=4.65.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0

# Optional: for enhanced functionality
pillow>=8.0.0
imageio>=2.9.0

# Development (optional)
pytest>=6.0.0
jupyter>=1.0.0